+++ app/Http/Controllers/UserController.php
@@ -8,6 +8,12 @@ use Spatie\Permission\Models\Role;
 
 class UserController extends Controller
 {
+    public function checkPermission() {
+        if(auth()->user()->hasRole('admin'))
+            return true;
+        else
+            return false;
+    }
     /**
      * Display a listing of the resource.
      *
@@ -15,6 +21,9 @@ class UserController extends Controller
      */
     public function index()
     {
+        if(!$this->checkPermission())
+            return redirect('home');
+
         $user = User::all();
         return view('admin.dashboard.user.user_list', compact('user'));
     }
@@ -48,6 +57,9 @@ class UserController extends Controller
      */
     public function show($id)
     {
+        if(!$this->checkPermission())
+            return redirect('home'); 
+        
         $user = User::find($id);
         return view('admin.dashboard.user.user_detail',compact('user'));
     }
@@ -60,12 +72,16 @@ class UserController extends Controller
      */
     public function edit($id)
     {
+        if(!$this->checkPermission())
+            return redirect('home');
+
         $role = Role::all();
         
         $role_name = [];
         foreach($role as $roles){
             $role_name[$roles->id] = $roles->name;
         }
+
         $user = User::find($id);
         return view('admin.dashboard.user.user_update',compact('user', 'role_name'));
     }
@@ -79,10 +95,19 @@ class UserController extends Controller
      */
     public function update(Request $request, $id)
     {
+        if(!$this->checkPermission())
+            return redirect('home');
+
         $this->validate($request,[ 
             'name' => 'required|string|max:250',
             'email' => 'required|string|max:250',
+            'role_id' => 'required',
         ]);
+
+        $user = User::find($id);
+        $role = Role::find($request->role_id);
+        $user->syncRoles([$role]);
+
         User::find($id)->update($request->all());
         return redirect()->route('user.index')->with('success','Record Updated Successfully');
     }
@@ -95,6 +120,9 @@ class UserController extends Controller
      */
     public function destroy($id)
     {
+        if(!$this->checkPermission())
+            return redirect('home');
+        
         User::find($id)->delete();
         return redirect()->route('user.index')->with('success','Record Deleted Successfully');
     }
